requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-registration-redux==1.6
django-social-share==1.1.2
django-summernote==0.8.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
boto3
django-storages